[EN] Start importer in a daemon, api as web, in new main.
This branch continues the migration from Immutant to Immutant 2.
The previous version ran `immutant.daemons/daemonize` with a :singleton?
false option. Unfortunately, Immutant 2 does not allow such an
option. It is always run as a singleton, which means that if run in a
wildfly cluster, at most one version will be running on the cluster at
the same time.

If we are not using a wildfly cluster, the point is moot and the
solution in this branch is satisfactory.

If we are using a wildfly cluster, we have to think about whether
running it as a singleton is reasonable, and if not, find a new
solution.

I have asked on IRC but I got no answer. Web searches turn up nothing.